move modules to /lib
